pyproject.toml
[project]
name = "ChurchSong"
version = "0.10.13"
authors = [{ name = "Stefan Bellon" }]
description = "Download the event agenda from ChurchTool and instantiate a PowerPoint slide template with the names and portraits of service staff as well as song database verification."
readme = "README.md"
license = "MIT"
license-files = ["LICENSE"]
requires-python = ">=3.14"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: Microsoft :: Windows",
    "Operating System :: POSIX :: Linux",
]
dependencies = [
    "click>=8.1.8",
    "packaging>=25.0",
    "platformdirs>=4.3.6",
    "polib>=1.2.0",
    "prettytable>=3.12.0",
    "psutil>=6.1.0; sys_platform == 'win32'",
    "pydantic>=2.9.2",
    "pypdf>=6.1.1",
    "python-pptx>=1.0.2",
    "reportlab>=4.4.4",
    "requests>=2.32.3",
    "rich>=14.0.0",
    "textual>=3.0.1",
    "typer>=0.15.4",
    "tzlocal>=5.3.1",
    "xlsxwriter>=3.2.3",
]

[project.urls]
Homepage = "https://github.com/sbellon/ChurchSong"
Repository = "https://github.com/sbellon/ChurchSong.git"
Changelog = "https://github.com/sbellon/ChurchSong/blob/main/CHANGELOG.md"

[project.scripts]
ChurchSong = "churchsong.__main__:main"

[dependency-groups]
dev = ["babel>=2.17.0"]
lint = ["pyright>=1.1.400", "ruff>=0.11.9"]

[build-system]
requires = ["uv_build>=0.7.19"]
build-backend = "uv_build"

[tool.uv]
package = true

[tool.setuptools.package-data]
"churchsong" = ["locales/*.po"]

[tool.pyright]
venvPath = "."
venv = ".venv"
typeCheckingMode = "strict"

[tool.ruff]
line-length = 88
builtins = ["_"]

[tool.ruff.lint]
select = ["ALL"]

fixable = [
    "I", # isort
]

ignore = [
    "D",      # docstring
    "COM812", # missing-trailing-comma
    "S113",   # request-without-timeout
    "S603",   # subprocess-without-shell-equals-true
    "FBT001", # boolean-type-hint-positional-argument
    "FBT002", # boolean-default-value-positional-argument
]

[tool.ruff.lint.flake8-quotes]
inline-quotes = "single"

[tool.ruff.format]
quote-style = "single"
line-ending = "lf"

[tool.ruff.lint.pep8-naming]
# Allow Pydantic's `@validator` decorator to trigger class method treatment.
classmethod-decorators = [
    "classmethod",
    "pydantic.validator",
    "pydantic.root_validator",
]

[tool.typos.default]
extend-ignore-re = [
    "(?Rm)^.*(#|//)\\s*typos:\\s*disable$",
    "(?s)(#|//)\\s*typos:\\s*off.*?\\n\\s*(#|//)\\s*typos:\\s*on",
]

[tool.typos.type.po]
extend-glob = ["*.po"]
check-file = false

[tool.bumpversion]
current_version = "0.10.13"
parse = "(?P<major>\\d+)\\.(?P<minor>\\d+)\\.(?P<patch>\\d+)"
serialize = ["{major}.{minor}.{patch}"]
search = "{current_version}"
replace = "{new_version}"
regex = false
ignore_missing_version = false
ignore_missing_files = false
tag = false
sign_tags = false
tag_name = "v{new_version}"
tag_message = "Bump version: {current_version} → {new_version}"
allow_dirty = false
commit = true
message = "Bump version: {current_version} → {new_version}"
moveable_tags = []
commit_args = ""
setup_hooks = []
pre_commit_hooks = ["uv sync", "git add uv.lock"]
post_commit_hooks = []

[[tool.bumpversion.files]]
filename = "pyproject.toml"

[[tool.bumpversion.files]]
filename = "CHANGELOG.md"
search = "Unreleased"
replace = "{new_version} ({now:%Y-%m-%d})"
